+++ sdk/databoxedge/arm-databoxedge/src/api/deviceCapacityCheck/index.ts
@@ -1,0 +1,5 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export { checkResourceCreationFeasibility } from "./operations.js";
+export type { DeviceCapacityCheckCheckResourceCreationFeasibilityOptionalParams } from "./options.js";