+++ sdk/domainservices/arm-domainservices/src/api/index.ts
@@ -1,0 +1,8 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export type {
+  DomainServicesResourceProviderContext,
+  DomainServicesResourceProviderOptionalParams,
+} from "./domainServicesResourceProviderContext.js";
+export { createDomainServicesResourceProvider } from "./domainServicesResourceProviderContext.js";