+++ sdk/netapp/arm-netapp/samples/v24/javascript/backupVaultsCreateOrUpdateSample.js
@@ -8,7 +8,7 @@ const { DefaultAzureCredential } = require("@azure/identity");
  * This sample demonstrates how to create or update the specified Backup Vault in the NetApp account
  *
  * @summary create or update the specified Backup Vault in the NetApp account
- * x-ms-original-file: 2026-01-01/BackupVaults_Create.json
+ * x-ms-original-file: 2026-04-01/BackupVaults_Create.json
  */
 async function backupVaultCreateOrUpdate() {
   const credential = new DefaultAzureCredential();