+++ sdk/netapp/arm-netapp/samples/v24/typescript/src/volumesReplicationStatusSample.ts
@@ -8,7 +8,7 @@ import { DefaultAzureCredential } from "@azure/identity";
  * This sample demonstrates how to get the status of the replication
  *
  * @summary get the status of the replication
- * x-ms-original-file: 2026-01-01/Volumes_ReplicationStatus.json
+ * x-ms-original-file: 2026-04-01/Volumes_ReplicationStatus.json
  */
 async function volumesReplicationStatus(): Promise<void> {
   const credential = new DefaultAzureCredential();