+++ sdk/postgresql/arm-postgresql-flexible/samples/v10-beta/typescript/src/nameAvailabilityCheckWithLocationSample.ts
@@ -8,7 +8,7 @@ import { DefaultAzureCredential } from "@azure/identity";
  * This sample demonstrates how to check the availability of name for resource
  *
  * @summary check the availability of name for resource
- * x-ms-original-file: 2026-01-01-preview/NameAvailabilityCheckWithLocation.json
+ * x-ms-original-file: 2026-04-01-preview/NameAvailabilityCheckWithLocation.json
  */
 async function checkTheValidityAndAvailabilityOfTheGivenNameInTheGivenLocationToAssignItToANewServerOrToUseItAsTheBaseNameOfANewPairOfVirtualEndpoints(): Promise<void> {
   const credential = new DefaultAzureCredential();