+++ sdk/web-pubsub/web-pubsub-client/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features Added
 
+- Added group streaming support
+
 ### Breaking Changes
 
 ### Bugs Fixed